Setting travis OTP-Versions
Setting the otp_release in `.travis.yml` to `R16B03`, `17.5`, and
`18.2`, as this are the most recent versions available on travis for the
last 3 major OTP-releases.

Personally I need 18+ for a small experimental project of my own keeping
compatibility to 16b3 for contests at my university until we get an
update there.